rename and map assets folder
